fix orderadd

+++ Fall2023_SWP391_NET1704_Vitvang/src/main/java/com/vitvang/productionmanagement/controller/AddOrderController.java
@@ -48,7 +48,7 @@ public class AddOrderController extends HttpServlet {
             if (totalprice == null) {
                   totalprice = "0";
             }
-            int total = Integer.parseInt(totalprice);
+            double total = Double.parseDouble(totalprice);
             String Address = request.getParameter("txtAddress");
 
             String[] CageID = request.getParameterValues("txtOrderCageID");
